\name{DUE-package}
\alias{DUE-package}
\alias{DUE}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab DUE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-06\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Roger Day <day01@pitt.edu>
}
\references{
}
\keyword{ package }
